refactor(shared-kernel): consolidate ApiResponse wrapper
