--- Backend/EY.HRPlatform.CoreHR/Models/Responses/ApiResponse.cs
@@ -1,20 +1,0 @@
-namespace EY.HRPlatform.CoreHR.Models.Responses;
-
-public class ApiResponse<T>
-{
-    public T? Data { get; set; }
-    public List<string> Errors { get; set; } = [];
-    public bool IsSuccess => Errors.Count == 0;
-
-    public static ApiResponse<T> Success(T data) => new() { Data = data };
-    public static ApiResponse<T> Failure(params string[] errors) => new() { Errors = [.. errors] };
-}
-
-public class ApiResponse
-{
-    public List<string> Errors { get; set; } = [];
-    public bool IsSuccess => Errors.Count == 0;
-
-    public static ApiResponse Success() => new();
-    public static ApiResponse Failure(params string[] errors) => new() { Errors = [.. errors] };
-}
